requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
psycopg2-binary==2.9.3
Werkzeug==2.0.3
